\name{cens1910}
\alias{cens1910}

\title{Census file for black literacy in 1910}

\description{
  Black literacy in 1910 for 1040 precincts. Rows are precincts and columns are:
  T (proportion of population over the age of 10 who can read)
  X (proportion of those who are black)
  1/N (where N is the number of people of age over 10)
  truthB (Truth for the proportion of literate black)
  truthW (Truth for the proportion of literate white)
}

\usage{data(cens1910)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 5 columns and 1040 rows, which
  is the input to the demos eip.cens1910 and einp.cens1910.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
