     \name{census1910}
     \docType{data}
     \alias{census1910}
     \title{Black Literacy in 1910}
     \description{
     	A dataset of aggregate literacy rates (t) and fraction of the population that is black (x), from the 1910 US Census.  Each observation represents one county.
     }
     \usage{census1910}
     \format{A data frame containing 1030 observations.}
     \source{Gary King, 1997, "Replication data for: A Solution to the Ecological Inference Problem: Reconstructing Individual Behavior from Aggregate Data", http://hdl.handle.net/1902.1/LWMMKUTYXS UNF:3:DRWozWd89+vNLO7lY2AHbg== IQSS Dataverse Network [Distributor] V3 [Version]}
     \references{
     Gary King (1997). \emph{A Solution to the Ecological Inference Problem.}  Princeton: Princeton University Press.  Section 13.2:241-5.
     Robinson, William S.  (1950)  ``Ecological Correlation and the Behavior of Individuals.''  \emph{American Sociological Review} 15:351-357.
       New York: Wiley.
     }
     \keyword{datasets}