% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sim.R
\name{ei.sim}
\alias{ei.sim}
\title{Simulate EI Solution via Importance Sampling}
\usage{
ei.sim(ei.object, ndraws = 99, nsims = 100)
}
\arguments{
\item{ei.object}{\code{ei} object}

\item{ndraws}{integer. The number of draws.}

\item{nsims}{integer. The number of simulations within each draw.}
}
\value{
`ei.sim` object
}
\description{
Simulate EI Solution via Importance Sampling
}
\examples{
data(sample_ei)
form <- t ~ x
ei_obj <- ei(form, total = "n", data = sample_ei, simulate = FALSE)
sims <- ei.sim(ei_obj)
}
\references{
Gary King (1997). A Solution to the Ecological Inference
Problem.  Princeton: Princeton University Press.
}
\author{
Gary King <<email: king@harvard.edu>> and Molly Roberts <<email:
molly.e.roberts@gmail.com>>
}
