% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{ei_sim}
\alias{ei_sim}
\title{Run Ecological Inference Simulation}
\usage{
ei_sim(data, ndraws = 99, nsims = 100)
}
\arguments{
\item{data}{an \code{ei_tbl} object from \code{ei_est()}}

\item{ndraws}{integer, default 99. The number of draws.}

\item{nsims}{integer, default 10. The number of simulations with each draw.}
}
\value{
ei_tbl
}
\description{
Run Ecological Inference Simulation
}
\examples{
data(sample_ei)
dbuf <- ei_est(sample_ei, x, t, n) \%>\% ei_sim()
}
\concept{tidy}
