% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_tbl.R
\name{as_ei_tbl}
\alias{as_ei_tbl}
\alias{ei_as_ei_tbl}
\title{Convert to \code{ei_tbl} objects}
\usage{
as_ei_tbl(x)

ei_as_ei_tbl(ei.object)
}
\arguments{
\item{x}{an object to be coerced}

\item{ei.object}{list-based ei object to convert to tibble-based object}
}
\value{
ei_tbl object
}
\description{
Convert to \code{ei_tbl} objects
}
\examples{
data(sample_ei)
form <- t ~ x
dbuf <- ei(form, total = "n", data = sample_ei)
dbuf <- ei_as_ei_tbl(dbuf)
}
\concept{tidy}
