\name{fultongen}
\alias{fultongen}

\title{Voter transitions in Fulton County
  in Georgia, 1994}

\description{
  Fulton County in Georgia voter transitions in 1994 for 289 precincts. Rows are precincts and columns are:\newline
  T (proportion of voting-age population turning out to vote in 1994)\newline
  X (proportion of voting-age population who turns out to vote in 1992)\newline
  N (the number of people of voting age in 1994)
}

\usage{data(fultongen)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 3 columns and 289 rows, which
  is the input to the demos eip.fultongen and einp.fultongen.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
