\name{hisp}
\alias{hisp}

\title{Hispanic voter in Pennsylvania's second state senate district 1990}   
  

\description{
  Hispanic who vote in the Pennsylvania's election, number of precinct
  is 252 and columns are:
  X the hispanic population
  T the turn-out of voter population
 
}

\usage{data(hisp)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 2 columns and 252 rows.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
