\name{in90}
\alias{in90}

\title{Marion County Indiana U.S. Senate election, 1990.}

\description{
  Marion County Indiana,1990, U.S. Senate election for 661
  precincts. Rows are precincts and columns are:
  X (proportion of voting-age population who are black)
  V (proportion of democratic votes)
  T (proportion of voting-age population turning out to vote)
  e2 (additional covariate)
  1/N (where N is the number of people of voting age)
 
}

\usage{data(in90)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 5 columns and 661 rows, which
  is the input to the demos eip.in90 and einp.in90.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
