\name{kyck88}
\alias{kyck88}

\title{Voter registration by race in Kentucky.}

\description{
  Voter registration by race in Kentucky for 118 precincts. Rows are precincts and columns are:
  T (proportion of registered voters for all population)
  X (proportion of voting-age population who are black)
  1/N (where N is the number of people of voting age)
  truthB (Truth for the proportion of registered black voters)
  truthW (Truth for the proportion of registered white voters)
}

\usage{data(kyck88)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 5 columns and 118 rows, which
  is the input to the demos eip.kyck88 and einp.kyck88.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
