\name{lavoteall}
\alias{lavoteall}

\title{Turnout vote by race in Loisiana in 1990.}

\description{
  Turnout vote by race in Loisiana in 1990 for 3262 precincts. Rows are precincts and columns are:\newline
  T (proportion of voting-age population turning out to vote)\newline
  X (proportion of voting-age population who are black)\newline
  N (N is the number of people of voting age)
}

\usage{data(lavoteall)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 3 columns and 3262 rows, which
  is the input to the demos eip.lavoteall and einp.lavoteall.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
