\name{nj}
\alias{nj}

\title{Non-minority turnout in New Jersey cities and towns in 1990}

\description{
  Non-minority turnout in New Jersey cities and towns in 1990 for 583
  precincts.Rows are precincts and columns are:\newline
  T (proportion of voting-age population turning out to vote)\newline
  X (proportion of voting-age population who are non-minority)\newline
  N (the number of people of voting age)
}

\usage{data(nj)}

\format{
  A fixed width \file{.txt} file.}
\details{Data file of numeric values with 3 columns and 583 rows, which
  is the input to the demos eip.nj and einp.nj.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
