\name{pa90}
\alias{pa90}

\title{Pennsylvania gubernatorial election in 1990}

\description{
  Pennsylvania gubernatorial election in 1990 for 1677 precincts. Rows
  are precincts and columns are:\newline
  X (proportion of voting-age population who are black)\newline
  V (proportion of democratic votes)\newline
  T (proportion of voting-age population turning out to vote)\newline
  e2 (additional covariate)\newline
  1/N (where N is the number of people of voting age)
 
}

\usage{data(pa90)}

\format{
  A fixed width \file{.txt} file.}
\details{Data file of numeric values with 5 columns and 1677 rows, which
  is the input to the demos eip.pa90 and einp.pa90.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
