% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_movie.R
\name{plot_movie}
\alias{plot_movie}
\title{Visualizing EI (Shiny visualization)}
\usage{
plot_movie(ei.object)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}
}
\description{
Visualizing EI (Shiny visualization)
}
\concept{visualization}
