% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sims.R
\name{plot_sims}
\alias{plot_sims}
\title{Visualizing EI (simulation)}
\usage{
plot_sims(ei.object)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\description{
Visualizing EI (simulation)
}
\concept{visualization}
