% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tomogRxC.R
\name{plot_tomogRxC}
\alias{plot_tomogRxC}
\title{Visualizing EI (tomography plot for RxC)}
\usage{
plot_tomogRxC(formula, data, total = NULL)
}
\arguments{
\item{formula}{A formula of the form \code{cbind(col1,
col2,...)~cbind(row1,row2,...)}}

\item{data}{data that contains the data that corresponds to the formula}

\item{total}{`total' is the name of the variable in the dataset that contains the number of individuals in each unit}
}
\value{
a ggplot object
}
\description{
A tomography plot for an estimated Ecological Inference model in RxC data.
}
\examples{
data(RxCdata)
formula <- cbind(turnout, noturnout) ~ cbind(white, black, hisp)
plot_tomogRxC(formula, RxCdata)
}
\concept{visualization}
