% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_truth.R
\name{plot_truth}
\alias{plot_truth}
\title{Visualizing EI (with truth)}
\usage{
plot_truth(ei.object)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}
}
\value{
a ggplot object
}
\description{
Compares truth to estimates at the district and precinct-level. Requires the \code{truth} argument in the \code{ei} object.
}
\examples{
data(matproii)
truth <- cbind(matproii$tb, matproii$tw)
suppressMessages({
  ei_res <- ei(formula = t ~ x, total = "n", truth = truth, data = matproii)
})
plot_truth(ei_res)
}
\concept{visualization}
