% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xt.R
\name{plot_xt}
\alias{plot_xt}
\title{Visualizing EI (xt-plot)}
\usage{
plot_xt(
  ei.object,
  options = list(density = TRUE, fit = TRUE, CI = 0.8, goodman = FALSE)
)
}
\arguments{
\item{ei.object}{The output of \code{ei()}.}

\item{options}{The list of options \itemize{
   \item \strong{density}: Showing density estimate.
   \item \strong{fit}: Showing \eqn{X_i} by \eqn{T_i} on the plot with estimated \eqn{E(T_i|X_i)}
   \item \strong{CI}: Showing a confidence interval with a specfied statistical level.
   \item \strong{goodman}: Goodman's regression.
}}
}
\value{
a ggplot object
}
\description{
\eqn{X_i} by \eqn{T_i} scatterplot with circles sized proportional to \eqn{N_i}.
}
\examples{
data(matproii)
suppressMessages({
  ei_res <- ei(formula = t ~ x, total = "n", data = matproii)
})
# `plot_xt()` function
plot_xt(ei_res)
# `plot_xt()` with options
plot_xt(ei_res, options = list(CI = 0.95, fit = FALSE, goodman = TRUE))
}
\concept{visualization}
