% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{RxCdata}
\alias{RxCdata}
\title{Sample Dataset}
\format{
A data frame containing 60 observations.
}
\source{
Source
}
\description{
Sample EI data for RxC with estimates for White, Black, and Hispanic groups.
}
\references{
Gary King (1997). \emph{A Solution to the Ecological Inference
Problem.} Princeton: Princeton University Press.
}
\concept{data}
\keyword{datasets}
