\name{sample}
\alias{sample}

\title{Turning out black and white vote}

\description{
  Rows are the 75 precincts and columns are:
  T (proportion of voting-age population turning out to vote)
  X (proportion of voting-age population who are black)
  N (number of people of voting age)

}

\usage{data(sample)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 3 columns and 75 rows, which
  is the input to the demos eip.sample and einp.sample.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
