\name{sample2}
\alias{sample2}

\title{Turning out black, white and democratic votes}

\description{
  Rows are the 99 precincts and 4 columns:\newline
  X (proportion of voting-age population who are black)\newline
  V (proportion of democratic votes)\newline
  T (proportion of voting-age population turning out to vote)\newline
  N (number of people of voting age)

}

\usage{data(sample2)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 4 columns and 99 rows to
  demonstrate ei2.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
