% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_ei}
\alias{sample_ei}
\title{Sample Data for Black Votes}
\format{
A data.frame containing 141 observations.
}
\source{
Sample data
}
\description{
A sample data.frame containing turnout, proportion Black, and total voters.
}
\references{
Gary King (1997). \emph{A Solution to the Ecological Inference
Problem.} Princeton: Princeton University Press.
}
\concept{data}
\keyword{datasets}
