\name{scsp}
\alias{scsp}

\title{Proverty status by sex in South Carolina in 1990}

\description{
  Proverty status by sex in South Carolina for 3187 precincts. Rows are precincts and columns are:
  T (proportion of people below provery level)
  X (proportion of male population below proverty)
  N (N is the number of people below proverty)
 
}

\usage{data(scsp)}

\format{
  A fixed width \file{.txt} file.}
\details{Data file of numeric values with 3 columns and 3187 rows, which
  is the input to the demos eip.scsp and einp.scsp.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
