% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei.R
\name{values_ei}
\alias{values_ei}
\title{Returning an element in the ei object}
\usage{
values_ei(object, name)
}
\arguments{
\item{object}{An \code{ei} object from the function \code{ei}.}

\item{name}{The name of the element to extract from the \code{ei} object.}
}
\description{
Returning an element in the ei object
}
