\name{ei.sim}
\alias{ei.sim}

\title{ 
	Simulate EI Solution via Importance Sampling
}
\description{
	Simulate EI solution via importance sampling
}
\usage{
ei.sim(ei.object)
}
\arguments{
	\item{ei.object}{\code{ei} object}
}
 
\references{
Gary King (1997). A Solution to the Ecological Inference Problem.  Princeton: Princeton University Press.
}
\author{
  Gary King <<email: king@harvard.edu>> and Molly Roberts <<email: roberts8@fas.harvard.edu>>
}

