% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei.R
\name{ei.sim}
\alias{ei.sim}
\title{Simulate EI Solution via Importance Sampling}
\usage{
ei.sim(ei.object)
}
\arguments{
\item{ei.object}{\code{ei} object}
}
\value{
TODO
}
\description{
Simulate EI solution via importance sampling
}
\examples{
# TODO
}
\references{
Gary King (1997). A Solution to the Ecological Inference
Problem.  Princeton: Princeton University Press.
}
\author{
Gary King <<email: king@harvard.edu>> and Molly Roberts <<email:
molly.e.roberts@gmail.com>>
}
