     \name{eiRxCsample}
     \docType{data}
     \alias{eiRxCsample}
     \title{A Sample Dataset}
     \description{
     	A description for this dataset
     }
     \usage{eiRxCsample}
     \format{A data frame containing 93 observations.}
     \source{Source}
     \references{
     Gary King (1997). \emph{A Solution to the Ecological Inference Problem.}  Princeton: Princeton University Press.
     }
     \keyword{datasets}