     \name{lavoteall}
     \docType{data}
     \alias{lavoteall}
     \title{Turnout by Race in Louisiana}
     \description{
     	The fraction of blacks registered voters (x) and fraction of voter turnout (t) in each Louisiana precinct, along with the true fraction of black turnout (tb) and non-black turnout (tw).
     }
     \usage{lavoteall}
     \format{A data frame containing 3262 observations.}
     \source{Gary King, 1997, "Replication data for: A Solution to the Ecological Inference Problem: Reconstructing Individual Behavior from Aggregate Data", http://hdl.handle.net/1902.1/LWMMKUTYXS UNF:3:DRWozWd89+vNLO7lY2AHbg== IQSS Dataverse Network [Distributor] V3 [Version]}
     \references{
     Gary King (1997). \emph{A Solution to the Ecological Inference Problem.}  Princeton: Princeton University Press.  Section 1.4:22-4.
     }
     \keyword{datasets}