\name{matproII}
\alias{matproII}

\title{ Voter registarion by race in the southern states Florida,
  Lousiana, and North and South Carolina.}

\description{
  Voter registarion by race in the states Florida, Lousiana, North
  Carolina, and South Carolina for 3275 precincts. Rows are precincts and columns are:\newline
  T (proportion of voting-age population registered)\newline
  X (proportion of voting-age population who are black)\newline
  1/N (where N is the number of people of voting age)\newline
  truthB (Truth for the proportion of voting age black registered)\newline
  truthW (Truth for the proportion of voting age white registered)
}

\usage{data(matproII)}

\format{
  A fixed width \file{.txt} file.}

\details{Data file of numeric values with 5 columns and 3275 rows, which
  is the input to the demos eip.matproII and einp.matproII.}
 
\source{\url{http://dvn.iq.harvard.edu/dvn/dv/king}}

\references{\url{http://gking.harvard.edu/projects/ei}}

\keyword{datasets}
