% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{matproii}
\alias{matproii}
\title{Voter Registration by Race in Southern States}
\format{
A data frame containing 268 observations.
}
\source{
Gary King, 1997, "Replication data for: A Solution to the Ecological
Inference Problem: Reconstructing Individual Behavior from Aggregate Data",
http://hdl.handle.net/1902.1/LWMMKUTYXS UNF:3:DRWozWd89+vNLO7lY2AHbg== IQSS
Dataverse Network [Distributor] V3 [Version]
}
\description{
Aggregate voter registration and fraction black, in counties in Florida,
Louisiana, North Carolina and South Carolina
}
\references{
Gary King (1997). \emph{A Solution to the Ecological Inference
Problem.} Princeton: Princeton University Press.  Chapter 10.
}
\concept{data}
\keyword{datasets}
