     \name{nj}
     \docType{data}
     \alias{nj}
     \title{Nonminority Turnout in New Jersey}
     \description{
     	A description for this dataset
     }
     \usage{nj}
     \format{A data frame containing 493 observations.}
     \source{Gary King, 1997, "Replication data for: A Solution to the Ecological Inference Problem: Reconstructing Individual Behavior from Aggregate Data", http://hdl.handle.net/1902.1/LWMMKUTYXS UNF:3:DRWozWd89+vNLO7lY2AHbg== IQSS Dataverse Network [Distributor] V3 [Version]}
     \references{
     Gary King (1997). \emph{A Solution to the Ecological Inference Problem.}  Princeton: Princeton University Press.  Section 1.4:24-5.
     }
     \keyword{datasets}