% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nj}
\alias{nj}
\title{Nonminority Turnout in New Jersey}
\format{
A data frame containing 493 observations.
}
\source{
Gary King, 1997, "Replication data for: A Solution to the Ecological
Inference Problem: Reconstructing Individual Behavior from Aggregate Data",
http://hdl.handle.net/1902.1/LWMMKUTYXS UNF:3:DRWozWd89+vNLO7lY2AHbg== IQSS
Dataverse Network [Distributor] V3 [Version]
}
\description{
A description for this dataset
}
\references{
Gary King (1997). \emph{A Solution to the Ecological Inference
Problem.} Princeton: Princeton University Press.  Section 1.4:24-5.
}
\concept{data}
\keyword{datasets}
