% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bound.R
\name{plot_bound}
\alias{plot_bound}
\title{Visualizing EI (bound)}
\usage{
plot_bound(ei.object, options = list())
}
\arguments{
\item{ei.object}{The output of \code{ei()} (it should be used with the \code{truth} argument)}

\item{options}{The list of options \itemize{
   \item \strong{parameter}: A parameter to plot. It takes either \code{betab} or \code{betaw}. This option is only for the 2x2 case.
}}
}
\value{
a ggplot object
}
\description{
Visualizing EI (bound)
}
\examples{
# 2x2
data(matproii)
truth <- cbind(matproii$tb, matproii$tw)
suppressMessages({
  ei_res <- ei(formula = t ~ x, total = "n", truth = truth, data = matproii)
})
plot_bound(ei_res, options = list(parameter = "betab"))
plot_bound(ei_res, options = list(parameter = "betaw"))
# RxC
data(RxCdata)
formula <- cbind(turnout, noturnout) ~ cbind(white, black, hisp)
suppressMessages({
  ei_resRxC <- ei(formula, data = RxCdata)
})
plot_bound(ei_resRxC)
}
\concept{visualization}
