% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bound.R
\name{plot_bound}
\alias{plot_bound}
\title{Visualizing EI (bound)}
\usage{
plot_bound(ei.object, options = list(parameter = "betab"))
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\value{
a ggplot object
}
\description{
Visualizing EI (bound)
}
\examples{
# 2x2
data(matproii)
truth <- cbind(matproii$tb, matproii$tw)
suppressMessages({
  ei_res <- ei(formula = t ~ x, total = "n", truth = truth, data = matproii)
})
plot_bound(ei_res, options = list(parameter = "betab"))
plot_bound(ei_res, options = list(parameter = "betaw"))
# RxC
data(RxCdata)
formula <- cbind(turnout, noturnout) ~ cbind(white, black, hisp)
suppressMessages({
  ei_resRxC <- ei(formula, data = RxCdata)
})
plot_bound(ei_resRxC, options = list(parameter = "betab"))
plot_bound(ei_resRxC, options = list(parameter = "betaw"))
}
\concept{visualization}
