% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bound.R
\name{plot_bound}
\alias{plot_bound}
\title{Visualizing EI}
\usage{
plot_bound(ei.object, options = list())
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\description{
Visualizing EI
}
