% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Visualizing EI (density)}
\usage{
plot_density(ei.object, options = list(parameter = "betab"))
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\value{
a ggplot object
}
\description{
Visualizing EI (density)
}
\concept{visualization}
