% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_movie.R
\name{plot_movie}
\alias{plot_movie}
\title{Visualizing EI}
\usage{
plot_movie(ei.object, options = list())
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\description{
Visualizing EI
}
