% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sims.R
\name{plot_sims}
\alias{plot_sims}
\title{Visualizing EI (simulation)}
\usage{
plot_sims(ei.object)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}
}
\value{
a ggplot object
}
\description{
Visualizing EI (simulation)
}
\examples{
data(matproii)
suppressMessages({
  ei_res <- ei(formula = t ~ x, total = "n", data = matproii)
})
plot_sims(ei_res)
}
\concept{visualization}
