% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tomog.R
\name{plot_tomog}
\alias{plot_tomog}
\title{Visualizing EI (tomography plot)}
\usage{
plot_tomog(
  ei.object,
  options = list(color = TRUE, category = 0, linecolor = "length", CI = NULL, points =
    FALSE, contour_ML = FALSE, contour_posterior = FALSE)
)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\value{
a ggplot object
}
\description{
Visualizing EI (tomography plot)
}
\examples{
data(matproii)
suppressMessages({
  ei_res <- ei(formula = t ~ x, total = "n", data = matproii)
})
plot_tomog(ei_res)
plot_tomog(ei_res, options = list(linecolor = "betab"))
plot_tomog(ei_res, options = list(linecolor = "betaw", category = 5))
plot_tomog(ei_res, options = list(points = FALSE, CI = 0.8))
}
\concept{visualization}
