% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_truth.R
\name{plot_truth}
\alias{plot_truth}
\title{Visualizing EI (with truth)}
\usage{
plot_truth(ei.object)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\description{
Visualizing EI (with truth)
}
\concept{visualization}
