% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xt.R
\name{plot_xt}
\alias{plot_xt}
\title{Visualizing EI (xt-plot)}
\usage{
plot_xt(
  ei.object,
  options = list(density = TRUE, fit = TRUE, CI = 0.8, goodman = FALSE)
)
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\description{
Visualizing EI (xt-plot)
}
\concept{visualization}
