% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xt.R
\name{plot_xt}
\alias{plot_xt}
\title{Visualizing EI}
\usage{
plot_xt(ei.object, options = list())
}
\arguments{
\item{ei.object}{The output of \code{ei()}}

\item{options}{The list of options}
}
\description{
Visualizing EI
}
