     \name{RxCdata}
     \docType{data}
     \alias{RxCdata}
     \title{Sample Dataset}
     \description{
     	A description for this dataset
     }
     \usage{RxCdata}
     \format{A data frame containing 60 observations.}
     \source{Source}
     \references{
     Gary King (1997). \emph{A Solution to the Ecological Inference Problem.}  Princeton: Princeton University Press.
     }
     \keyword{datasets}