     \name{sample}
     \docType{data}
     \alias{sample}
     \title{Sample Data for Black Votes}
     \description{
     	A description for this dataset
     }
     \usage{sample}
     \format{A vector containing 141 observations.}
     \source{Source}
     \references{
     Gary King (1997). \emph{A Solution to the Ecological Inference Problem.}  Princeton: Princeton University Press.
     }
     \keyword{datasets}