% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample}
\alias{sample}
\title{Sample Data for Black Votes}
\format{
A vector containing 141 observations.
}
\source{
Source
}
\description{
A description for this dataset
}
\references{
Gary King (1997). \emph{A Solution to the Ecological Inference
Problem.} Princeton: Princeton University Press.
}
\concept{data}
\keyword{datasets}
