% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ei}
\alias{summary.ei}
\title{Summarize Ecological Inference Estimates}
\usage{
\method{summary}{ei}(object, ...)
}
\arguments{
\item{object}{An \code{ei} object from the function \code{ei}.}

\item{\dots}{A list of options to return in graphs.  See values below.}
}
\value{
TODO
}
\description{
`summary' method for the class `ei'.
}
\examples{

data(sample)
formula <- t ~ x
dbuf <- ei(formula = formula, total = "n", data = sample)
print(summary(dbuf))
}
\references{
Gary King (1997). A Solution to the Ecological Inference
Problem.  Princeton: Princeton University Press.
}
\author{
Gary King <<email: king@harvard.edu>> and Molly Roberts <<email:
molly.e.roberts@gmail.com>>
}
